backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
PyPDF2
python-docx
pdfplumber